notebooks/requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning
scikit-learn>=1.0.0
umap-learn>=0.5.2
joblib>=1.2.0
nltk>=3.8.1

# Visualization
plotly>=5.3.0
wordcloud>=1.8.0
plotnine>=0.8.0
bokeh>=2.4.0

# Jupyter & Development
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.5
jupyterlab>=3.0.0

# Data Processing
tqdm>=4.62.0
pandarallel>=1.5.4

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Code Quality
black>=21.12b0
flake8>=4.0.1
isort>=5.10.1
mypy>=0.910
